refactor(forge/llm): Create `BaseOpenAI(Chat|Embedding)Provider` and deduplicate `GroqProvider` & `OpenAIProvider`
- Add `_BaseOpenAIProvider`, `BaseOpenAIChatProvider`, and `BaseOpenAIEmbeddingProvider`
- (Re)move as much code as possible from `GroqProvider` and `OpenAIProvider`

Also:
- Rename `get_available_models()` to `get_available_chat_models()` on `BaseChatModelProvider`
- Add `get_available_models()` to `BaseModelProvider`
- Add `get_available_embedding_models()` to `BaseEmbeddingModelProvider`
- Move common `fix_failed_parse_tries` config attribute into base `ModelProviderConfiguration`
