+++ autogpt/tests/unit/test_config.py
@@ -144,7 +144,7 @@ def test_azure_config(config_with_azure: Config) -> None:
 @pytest.mark.asyncio
 async def test_create_config_gpt4only(config: Config) -> None:
     with mock.patch(
-        "forge.llm.providers.multi.MultiProvider.get_available_models"
+        "forge.llm.providers.multi.MultiProvider.get_available_chat_models"
     ) as mock_get_models:
         mock_get_models.return_value = [
             ChatModelInfo(
@@ -164,7 +164,7 @@ async def test_create_config_gpt4only(config: Config) -> None:
 @pytest.mark.asyncio
 async def test_create_config_gpt3only(config: Config) -> None:
     with mock.patch(
-        "forge.llm.providers.multi.MultiProvider.get_available_models"
+        "forge.llm.providers.multi.MultiProvider.get_available_chat_models"
     ) as mock_get_models:
         mock_get_models.return_value = [
             ChatModelInfo(
